% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\docType{class}
\name{GeneHancerDB-class}
\alias{GeneHancerDB-class}
\alias{.GeneHancerDB}
\alias{GeneHancerDB}
\title{GeneHancerDB}
\usage{
GeneHancerDB()
}
\value{
An object of the GeneHancerDB class
}
\description{
GeneHancerDB

Create a GeneHancerDB connection
}
