% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\name{getEnhancers,GeneHancerDB-method}
\alias{getEnhancers,GeneHancerDB-method}
\alias{getEnhancers}
\title{Get all the enhancer regions for the gene}
\usage{
\S4method{getEnhancers}{GeneHancerDB}(obj, targetGene, tissues = "all", maxSize = 10000)
}
\arguments{
\item{obj}{An object of class GeneHancerDB}

\item{targetGene}{default NA, in which case the current object's targetGene is used.}
}
\description{
Get all the enhancer regions for the gene
}
\seealso{
setTargetGene
}
