% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\name{getEnhancerTissues,GeneHancerDB-method}
\alias{getEnhancerTissues,GeneHancerDB-method}
\alias{getEnhancerTissues}
\title{Get all the enhancer tissues included in the current genehancer}
\usage{
\S4method{getEnhancerTissues}{GeneHancerDB}(obj, targetGene)
}
\arguments{
\item{obj}{An object of class GeneHancerDB}

\item{targetGene}{a character string}
}
\description{
Get all the enhancer tissues included in the current genehancer
}
\seealso{
getEnhancers
}
