% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\name{listTissues,GeneHancerDB-method}
\alias{listTissues,GeneHancerDB-method}
\alias{listTissues}
\title{Return a character vector containing all of the tissues known to GeneHancer}
\usage{
\S4method{listTissues}{GeneHancerDB}(obj, targetGene = NA)
}
\arguments{
\item{obj}{An object of class GeneHancerDB}

\item{targetGene}{A character string, default NA in which case all tissues for all genes are returned}
}
\description{
Return a character vector containing all of the tissues known to GeneHancer
}
