% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\name{queryByRegion,GeneHancerDB-method}
\alias{queryByRegion,GeneHancerDB-method}
\alias{queryByRegion}
\title{Get all the enhancer regions for the gene}
\usage{
\S4method{queryByRegion}{GeneHancerDB}(obj, chrom, start, end)
}
\arguments{
\item{obj}{An object of class GeneHancerDB}

\item{chrom}{character, the chromosome name WITHOUT leading 'chr'}

\item{start}{numeric the starting base}

\item{end}{numeric the ending base}
}
\description{
Get all the enhancer regions for the gene
}
