% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\name{retrieveEnhancersFromDatabase,GeneHancerDB-method}
\alias{retrieveEnhancersFromDatabase,GeneHancerDB-method}
\alias{retrieveEnhancersFromDatabase}
\title{Get all the enhancer regions for the gene}
\usage{
\S4method{retrieveEnhancersFromDatabase}{GeneHancerDB}(obj, targetGene, tissues)
}
\arguments{
\item{obj}{An object class GeneHancerDB}

\item{targetGene}{a HUGO gene symbol}

\item{tissues}{"all" or a vector of case-agnostic tissue names}
}
\description{
Get all the enhancer regions for the gene
}
\seealso{
listTissues
}
