% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\name{to.hg19,GeneHancerDB-method}
\alias{to.hg19,GeneHancerDB-method}
\alias{to.hg19}
\title{hg38 coordinates are default, but sometimes we need hg19.}
\usage{
\S4method{to.hg19}{GeneHancerDB}(obj, tbl)
}
\arguments{
\item{obj}{An object of class GeneHancerDB}

\item{data.frame}{with (at least) chrom, start, end}
}
\description{
hg38 coordinates are default, but sometimes we need hg19.
}
