% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\docType{data}
\name{GeneHancerDB-class}
\alias{GeneHancerDB-class}
\alias{GENEHANCER_VERSION}
\alias{GeneHancerDB}
\title{GeneHancerDB}
\format{
An object of class \code{character} of length 1.
}
\usage{
GENEHANCER_VERSION

GeneHancerDB()
}
\value{
An object of the GeneHancerDB class
}
\description{
GeneHancerDB

Create a GeneHancerDB connection
}
\keyword{datasets}
